desain_intj

+++ modules/produk/hapus.php
@@ -2,16 +2,27 @@
 
 include '../../config/koneksi.php';
 
-$id = $_GET['id'];
-
-$data = mysqli_fetch_assoc(mysqli_query($conn,
-"SELECT * FROM produk WHERE id_produk='$id'"));
-
-unlink('../../assets/upload/'.$data['gambar']);
-
-mysqli_query($conn,
-"DELETE FROM produk WHERE id_produk='$id'");
-
-header("Location: index.php");
+if(isset($_GET['id'])) {
+    $id = $_GET['id'];
+    
+    // Ambil data gambar dulu
+    $query = mysqli_query($conn, "SELECT gambar FROM produk WHERE id_produk='$id'");
+    $data = mysqli_fetch_assoc($query);
+    
+    // Hapus file gambar jika ada
+    if($data['gambar'] != '' && file_exists('../../assets/upload/'.$data['gambar'])) {
+        unlink('../../assets/upload/'.$data['gambar']);
+    }
+    
+    // Hapus data dari database
+    mysqli_query($conn, "DELETE FROM produk WHERE id_produk='$id'");
+    
+    echo "<script>
+            alert('Produk berhasil dihapus');
+            document.location.href = 'index.php';
+          </script>";
+} else {
+    header("Location: index.php");
+}
 
 ?>
